requirements.txt
torch
torchvision
tensorboard
scipy
numpy
matplotlib
opencv_python
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
